moved todo logic to context provider
